separated data classes into domain and data layers
