+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/app/navDto/dto/GameCreatedNavDto.kt
@@ -1,0 +1,11 @@
+package pw.checkers.app.navDto.dto
+
+import kotlinx.serialization.Serializable
+import pw.checkers.game.domain.model.PlayerColor
+
+@Serializable
+data class GameCreatedNavDto(
+    val gameId: String,
+    val color: PlayerColor,
+    val opponent: UserNavDto,
+)
