+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/app/navDto/dto/JoinedQueueNavDto.kt
@@ -1,0 +1,8 @@
+package pw.checkers.app.navDto.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class JoinedQueueNavDto(
+    val message: String
+)
