--- frontend/composeApp/src/commonMain/kotlin/pw/checkers/data/domain/Result.kt
@@ -1,16 +1,0 @@
-package pw.checkers.data.domain
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-enum class Result {
-    @SerialName("draw")
-    DRAW,
-
-    @SerialName("white")
-    WHITE,
-
-    @SerialName("black")
-    BLACK,
-}
