--- frontend/composeApp/src/commonMain/kotlin/pw/checkers/data/request/JoinQueue.kt
@@ -1,10 +1,0 @@
-package pw.checkers.data.request
-
-import kotlinx.serialization.Serializable
-import pw.checkers.data.Content
-import pw.checkers.data.domain.User
-
-@Serializable
-data class JoinQueue(
-    val user: User
-) : Content
