--- frontend/composeApp/src/commonMain/kotlin/pw/checkers/data/response/GameEnd.kt
@@ -1,10 +1,0 @@
-package pw.checkers.data.response
-
-import kotlinx.serialization.Serializable
-import pw.checkers.data.Content
-import pw.checkers.data.domain.Result
-
-@Serializable
-data class GameEnd(
-    val result: Result
-) : Content
