--- frontend/composeApp/src/commonMain/kotlin/pw/checkers/data/response/Possibilities.kt
@@ -1,10 +1,0 @@
-package pw.checkers.data.response
-
-import kotlinx.serialization.Serializable
-import pw.checkers.data.Content
-import pw.checkers.data.domain.Cell
-
-@Serializable
-data class Possibilities(
-    val moves: List<Cell>,
-) : Content
