+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/game/data/dto/CellDto.kt
@@ -1,0 +1,9 @@
+package pw.checkers.game.data.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class CellDto (
+    val row: Int,
+    val col: Int,
+)
