+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/game/data/dto/PlayerColorDto.kt
@@ -1,0 +1,10 @@
+package pw.checkers.game.data.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class PlayerColorDto() {
+    @SerialName("white") WHITE,
+    @SerialName("black") BLACK
+}
