+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/game/data/dto/ResultDto.kt
@@ -1,0 +1,11 @@
+package pw.checkers.game.data.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class ResultDto {
+    @SerialName("draw") DRAW,
+    @SerialName("white") WHITE,
+    @SerialName("black") BLACK
+}
