+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/game/domain/model/Cell.kt
@@ -1,0 +1,7 @@
+package pw.checkers.game.domain.model
+
+data class Cell (
+    val row: Int,
+    val col: Int,
+    val piece: Piece? = null,
+)
