+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/game/domain/model/Move.kt
@@ -1,0 +1,8 @@
+package pw.checkers.game.domain.model
+
+data class Move(
+    val fromRow: Int,
+    val fromCol: Int,
+    val toRow: Int,
+    val toCol: Int,
+)
