+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/game/domain/model/Piece.kt
@@ -1,12 +1,8 @@
-package pw.checkers.data.domain
+package pw.checkers.game.domain.model
 
-import kotlinx.serialization.Serializable
-
-@Serializable
 enum class PieceType() {
     PAWN, QUEEN
 }
 
-@Serializable
 data class Piece(val color: PlayerColor, val type: PieceType)
