+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/game/domain/model/Result.kt
@@ -1,0 +1,7 @@
+package pw.checkers.game.domain.model
+
+enum class Result {
+    DRAW,
+    WHITE,
+    BLACK,
+}
