+++ frontend/composeApp/src/commonMain/kotlin/pw/checkers/game/domain/model/User.kt
@@ -1,0 +1,5 @@
+package pw.checkers.game.domain.model
+
+data class User(
+    val username: String
+)
